requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
coc.py==0.3.2
discord.py==1.2.5
idna==2.8
idna-ssl==1.1.0
multidict==4.7.3
pycparser==2.19
PyNaCl==1.3.0
six==1.13.0
typing-extensions==3.7.4.1
websockets==6.0
yarl==1.4.2
